[Fix] GameCameraView Hero위로 움직이도록 수정 [Fix] EnemySpawn 랜덤생성 [Add] Round별로 Enemy난이도 수정 [Add] Lose시에는 Round 진행사항 반영 안됨

+++ Project_G/Assets/_BKT/Scripts/Objects/Creature/Enemy/Enemy.cs
@@ -7,7 +7,7 @@ using static Define;
 public class Enemy : Creature
 {
     protected IDamageable Target { get; private set; }
-    public int Power { get; protected set; }
+    public float Power { get; protected set; }
     public Define.EColorType ColorType { get; protected set; }
     protected UI_WorldSpace_Hp UI_EnemyHp { get; set; }
     private bool TriggerBuilding { get; set; }
